relacionamentos organizados e testados

+++ cheku/src/main/java/com/cheku/cheku/api/APIDeleteController.java
@@ -13,8 +13,27 @@ public class APIDeleteController {
     @Autowired
     private CarService carService;
 
-    @DeleteMapping("api/car/{car_id}")
-    public String  deleteCar(@PathVariable Long car_id) {
-         return carService.deleteCar(car_id);
+    @Autowired
+    private GroupService groupService;
+
+    @DeleteMapping("api/user/{user_id}/group/{group_id}/car/{car_id}")
+    public String  deleteCar(@PathVariable Long user_id, @PathVariable Long group_id, @PathVariable Long car_id) {
+        //verificar se o user_id é o admin do grupo
+        if (groupService.verifyAdmin(user_id, group_id)) {
+            carService.deleteCar(car_id);
+            return "Carro apagado com sucesso";
+        }
+        return "Não tem permissões para apagar o carro";
+    }
+
+    @DeleteMapping("api/user/{user_id}/group/{group_id}")
+    public void  deleteCar(@PathVariable Long user_id, @PathVariable Long group_id) throws RuntimeException {
+        //verificar se o user_id é o admin do grupo
+        if (groupService.verifyAdmin(user_id, group_id)) {
+            groupService.deleteGroup(group_id);
+        }
+        else{
+            throw new RuntimeException("Não tem permissões para apagar o grupo");
+        }
     }
 }
